Refactor questions a bit for better coherence + remove unnecessary index in approval votings
refs #886
